pyproject.toml
[project]
name = "y-finance-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow==2.8.1",
]
